Commit initial files and test files
Started development of the Flask app for hosting the site and started
working on the html for the headers and footers. Json files containting
examples of the entries in the Mongodb have also been added. While the
project is still in early development these files are very much subject
to change.