Add base,login,recipeform pages, js and python
Some of the early building blocks of the site have been added in this
commit. The base html has been better implemented with relevant css and
icon files. Some new JavaScript libraries have been included where
needed. The html files for the basic structure of the login and recipe
submission pages have also been added. A JavaScript file has been added
for those pages that require it. The code to pull the data from the
recipe form has been added to the flask file. The data is formatted
correctly according to Schema standards and is ready to be pushed to
MongoDB.