Add image picker to form and login/out function
So this is actually the contents of two commits, don't ask me how but I
accidentally added the files of what was to be the next commit to
this one so I'm splitting the message up into two:

Commit 1:
A modal was added to the recipe form so that the user can pick an image
from a Google image search. The code to handle the search and to resize
any images to fit in the square frames of the modal has been added as
well as any supporting CSS. Default images have also been added to be
used if the user does not want to choose an image.

It doesn't make sense to have template specific JavaScript to load in
the base template so any specific js has been moved to its corresponding
template. The issue with structuring things this way is that any
supporting js such as jQuery needs to be loaded before the template
meaning some js must be loaded at the top of the base page rather than
the bottom, tests need to be done later in development to see if this is
makes things better or not.

Lastly, python, js, HTML and CSS files are now automatically formatted
using Black for python and Prettier for the rest.

Commit 2:
The login and logout pages are now working correctly and use
user details from the mongodb rather than the hard coded credential from
before. A signup form has been added to add user details to the database
including hashed user passwords so they are not stored in plain text.