Add welcome page and draw data from list tables
Now the select2 lists on the recipe submission form pull data from list
tables in the mongodb so that users can select pre-existing ingredients,
units, etc. Since this is the first version of the site that will be
hosted on Heroku to be sent out to beta testers, a welcome page has been
added to explain the state of the project. It lists what parts of the
project are working and what parts are left to implement/fix. This is
only a temporary solution till the project is in a full functioning
state. Also in this commit all the links to help I have used for this
project have been added to the README, the +70 tabs I was keeping open
was getting messy.