Add more views and change structure of user page
Added functionality to the favorites and comments views on the user
page. They only uses a basic layout so some more styling will be needed,
but they're functionally there. The structure of the user page has also
been changed, the layout of the small recipe card now use a bootstrap
grid rather than a table and can be found in it's own template. The
comment card uses a very similar view and for now uses its own template
which has also been added. Changes to the flask app and CSS have been
made where needed.