Implement user info update form and fix bugs
Added a form to the user page to allow them to update their email or
password. The route to deal with pushing the updates to the Mongo
database has been added to the main flask app. The user route has now
been made dynamic, rather than being found at /user the users profile
url is now at /<username>. At the moment if the user tries to go to a
/<username> that is not their own then they are redirected to log in as
that user. The recipe cards for both the favorites and comments now use
one template rather than two so the duplicate template has been removed.
A bug with font awesome has been fixed where the star icons defined in
CSS would not render on mobile so now on smaller screen sizes a Unicode
icon is used instead.