Fix user page bug, change variable, add infocard
Forgot to change the template of the comments view to use the new merged
template so fixed that. Added an info card to the user page which gives
the user some information about their profile. One piece of information
included is the date the user created their profile so now a date field
is included when the user signs up. Changed some variable names so that
they are more consistent and unique.