Add recipe page with route and other bug fixes
Added the first near functional recipe page with flask route. The page
features recipe details pulled from the database and the ability for a
user to like and post comments on a page. When a user interacts with a
page that information is now added to the relevant entries in the
database. Some changes have been made to the user page, mainly replacing
some lists with tables to create a better layout. All relevant CSS
styles have been added as well.