Add delete and edit functions, add/move modals
Re-purposed the recipe submission form to allow the user to edit their
previously submitted recipes. A new route has been added to the flask
app and changes were made to the recipeform template and form.js to make
this work. The user can now also delete a submitted recipe, a
conformation modal is used to make sure the user is clear before the
recipe is deleted as at the moment deleted recipes cannot be recovered.
The star rating modals have been moved into their own templates so that
now all modals are stored in separate templates. Some serious testing
and refactoring needs to be done but at this point all 4 CRUD operations
have now been added to the project!