Fix delete recipe function, clean/comment code
The delete recipe function was not complete as it only deleted the recipe entry itself leaving ghosts behind to cause trouble. Like for example when loading the users favorites the page would look for the now non-existent recipe. Now when the recipe is deleted it is also removed from users favorites list, all the comments made on the recipe are removed from the comment database and the users track of those comments are also removed. Some of the templates are unlikely to be changed at this stage so they have been formatted using Prettier and some better comments have been added.