Implement home page with supporting templates
The home page for the site has been added with the majority of
functionality. The desktop style has been built with the mobile layout
to be further tweaked. There is an issue with the way images have been
implemented so that will be fixed before continuing on design work. The
Infinite Scrolling JavaScript library has been added to the site so that
recipes infinitely scroll on the home page. A new collection has been
added to the database which documents activity on the site which is then
displayed in a card on the side of the desktop view of the site. In some
files the code has been cleaned up so minor changes have been made.