Create search page and start accepting results
The basic layout of the search results page has been completed including
select2 boxes for advanced searching. The routes for accepting the
search terms has been added as well as JavaScript for the select2 boxes
and infinite scroll libraries. The code to preform the searches has yet
to be added. Also, most of the code has been recycled from other parts
of the site so it will need a once over to make sure that all irrelevant
code has been removed.