Add Contact Page and fix some spelling
Added the contact page that includes a contact form and a changelog. The
site is getting close to sending it out to many people so the changelog
is there to let people know what might have changed since they last
interacted with the site. Some minor spelling mistakes were also fixed
in some templates.