Implement email backend and escape all inputs
The email backend code has been added so that now the contact form sends
an email in the development environment and hopefully also in the live
environment. The config variables have been added to the local config
file and to the live Heroku environment. Some basic security has been
added to the user inputs were they are now escaped so that should
mitigate simple XSS but more needs to be done to mitigate more advanced
attacks.