Migrate database and add coming soon popups
Moved to a fresh new database from the previous test database. Added a
new element to a users entry so that in the future, users votes are
tracked so they can't vote multiple times for the same recipe. Coming
soon dialog boxes have also been added to some element that have not
been finished yet to let the user know that functionality is coming
soon.