Fix bugs in select2 boxes and improve scroll look
Removed the options that made removing the last selection from a select2
box fiddly and fixed an issue where the dialogue boxes would render over
the navbar at the top of the page. The dialogue boxes still appear above
when the select2 is opened near the bottom of the page but to change
that would require editing the source of the select2 library so I'll
leave it for the moment. Added the jQuery Scrollbar library to the
project to make the scrollbar a little nicer looking in the select2
boxes.