Fix bugs with comments on the home page
Fixed issues with the comments being displayed on the home page. The
comments are now passed into the recipecardlarge template as a list much
like how they are handled on the individual recipe page. This is handled
with a loop in the app.py and may not scale as the number of recipes
gets large but will do for the moment. An issue with the comments button
on the home page was also fixed. Before, when the button was pressed all
comments fields were reveled, now the recipecardlarge uses the recipes
id for the field so that when the button is pressed on the relevant
field is expanded.