Fix bug when rating users recipes
Fixed a bug which would allow a user to vote multiple times on a recipe.
Now every time a user adds their vote, the vote is recorded to their
profile. Next time the user rates the same recipe the profile is checked
and since the user has already voted a dialogue box pops up to tell them
that they have already voted. The system at the moment does not allow
for the user to amend their vote but significant work would need to be
done to allow that so it will be left for a future date.