Improve mobile layout and fix some bugs
Changed the way the advanced options are displayed on mobile, they are
now included in an accordion style dropdown that only appears on mobile.
The advanced layout has been moved to its own template as it needs to be
duplicated for the two layouts. Also some bugs in how the search results
are displayed have hopefully been fixed. A lot more recipes will need to
be added to the database in order to test everything works properly.