Cleanup comments, add wireframes + README report
The first draft of the README report has been added in this commit as
well as the wireframes for the project. Also there has been a major
cleanup of the comments and code in almost all files in the project.
Redundant, legacy and commented out code has been removed with remaining
code formatted using Prettier and Black. Comments have been better
formatted and expanded on where needed and reduced where not. Anything
other major changes are listed below:
 - Fix liked recipes not showing on the home page correctly
 - Inputs are now escaped to prevent XSS attacks
 - Comments related to Jinja code now use Jinja comment format
 - Random print and console.log outputs have been removed