Fix star rating, add errors, update changelog
Fixed the same issue with rendering the Font Awesome star icons as
before, this time in the rating modal CSS. Fixed a minor visual
quirk on the Contact page where the alert card would show even if there
wasn't an alert to be displayed. Updated the README with a list of
remaining errors/warnings from the linters as there is a known issue
caused by a Jinja variable.