Add Select2 sort, repeat homepage list, fix modal
Implemented the sorting of the Select2 options by usage number. Since
the number of recipes in the database is quite low, the list of recipes
on the home page now load twice to better test the endless scrolling.
Fixed an issue with the modal where it would appear behind the
'fixed-top' navbar. Also added a screenshot to the README.